refactor to support install via NPM package
