--- Plugins/iac-support-backend/catalog-info.yaml
@@ -1,12 +1,0 @@
-apiVersion: backstage.io/v1alpha1
-kind: Resource Component
-metadata:
-  name: backstage-plugin-catalog-backend-module-iacsupport-entity-model
-  title: '@backstage/plugin-catalog-backend-module-iacsupport-entity-model'
-  description: >-
-    Adds support for the IaC Support specific entity model (e.g. the Template
-    kind) to the catalog backend plugin.
-spec:
-  lifecycle: experimental
-  type: backstage-backend-plugin-module
-  owner: maintainers
