--- Plugins/iac-support-backend/src/processor/index.ts
@@ -1,1 +1,0 @@
-export { IacSupportEntitiesProcessor } from './IacSupportEntitiesProcessor';
