+++ plugins/iac-support-backend/README.md
@@ -1,3 +1,3 @@
-# @backstage/plugin-catalog-backend-module-scaffolder-entity-model
+# @paulpogo/plugin-iac-support-backend
 
 Adds support for the scaffolder specific entity model (e.g. the Template kind) to the catalog backend plugin.
