+++ plugins/iac-support-common/package.json
@@ -1,15 +1,26 @@
 {
-  "name": "@internal/plugin-iac-support-common",
+  "name": "@paulpogo/plugin-iac-support-common",
   "description": "Common functionalities for the iac-support plugin",
-  "version": "0.1.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
+  "version": "0.1.6",
+  "license": "MIT",
+  "main": "index.ts",
+  "types": "index.ts",
   "publishConfig": {
     "access": "public",
     "main": "dist/index.cjs.js",
     "module": "dist/index.esm.js",
     "types": "dist/index.d.ts"
+  },
+  "exports": {
+    ".": "./index.ts",
+    "./package.json": "./package.json"
+  },
+  "typesVersions": {
+    "*": {
+      "package.json": [
+        "package.json"
+      ]
+    }
   },
   "backstage": {
     "role": "common-library"
@@ -17,6 +28,8 @@
   "sideEffects": false,
   "scripts": {
     "build": "backstage-cli package build",
+    "tsc": "tsc",
+    "tsc:full": "tsc --skipLibCheck false --incremental false",
     "lint": "backstage-cli package lint",
     "test": "backstage-cli package test",
     "clean": "backstage-cli package clean",
@@ -27,7 +40,9 @@
     "@backstage/cli": "^0.25.1"
   },
   "files": [
-    "dist"
+      "dist",
+      "dist-types",
+      "config.d.ts"
   ],
   "dependencies": {
     "@backstage/catalog-model": "^1.4.3",
