+++ plugins/iac-support-common/src/EnvironmentEntity.ts
