+++ plugins/iac-support-common/src/ResourceComponentEntity.ts
