+++ plugins/iac-support-common/src/Schema/Environment.v1alpha1.schema.json
