+++ plugins/iac-support-common/src/Schema/ResourceComponent.v1alpha1.schema.json
